% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.data.frame.environment}
\alias{as.data.frame.environment}
\title{Coerce an environment to a data.frame}
\usage{
\method{as.data.frame}{environment}(x)
}
\arguments{
\item{x}{An R environment.}
}
\value{
A data.frame version of \code{x}
}
\description{
Coerce an environment to a data.frame
}
\examples{
iris_env <- as.environment(iris)
iris_df <- as.data.frame(iris_env)
identical(iris, mtcars_df)
}
