% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as_tibble.environment}
\alias{as_tibble.environment}
\title{Coerce a data frame to a tibble}
\usage{
\method{as_tibble}{environment}(x)
}
\arguments{
\item{x}{An environment}
}
\value{
A tibble version of \code{x}.
}
\description{
Coerce a data frame to a tibble
}
\examples{
iris_env <- as.environment(iris)
iris_tbl <- tibble::as_tibble(iris_env)
all(iris == iris_tbl)
}
