% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter.environment}
\alias{filter.environment}
\title{Keep rows that match a condition}
\usage{
\method{filter}{environment}(.data, ...)
}
\arguments{
\item{.data}{A tabular environment}

\item{...}{<\code{\link[rlang:args_data_masking]{data-masking}}> Expressions that
return a logical value, and are defined in terms of the variables in
\code{.data}. If multiple expressions are included, they are combined with the
\code{&} operator. Only rows for which all conditions evaluate to \code{TRUE} are
kept.}
}
\value{
An environment
}
\description{
Keep rows that match a condition
}
