% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glimpse.R
\name{glimpse.environment}
\alias{glimpse.environment}
\title{Get a glimpse of your data}
\usage{
\method{glimpse}{environment}(x, width = NULL)
}
\arguments{
\item{x}{An R environment}

\item{width}{Width of output: defaults to the setting of the
\code{width} \link[pillar:pillar_options]{option} (if finite)
or the width of the console.}
}
\value{
Original \code{x} is (invisibly) returned
}
\description{
Get a glimpse of your data
}
