% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.environment}
\alias{print.environment}
\title{Print values of an environment}
\usage{
\method{print}{environment}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Returns \code{x} invisibly
}
\description{
Print values of an environment
}
