% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tabular.R
\name{is_tabular}
\alias{is_tabular}
\title{Confirm an object is table shaped}
\usage{
is_tabular(x)
}
\arguments{
\item{x}{List or environment object to be checked.}
}
\value{
Logical. Whether all vectors in \code{x} have the same length.
}
\description{
Check whether every vector in a \link{list} or \link{environment} has the same
\code{\link[=length]{length()}}.
}
\examples{
# TRUE
is_tabular(iris)
is_tabular(as.environment(iris))
is_tabular(new.env())
is_tabular(as.environment(list(chr = letters, num = 1:26)))

# FALSE
is_tabular(list(letters, 1:3))
is_tabular(as.environment(list(chr = letters, num = 1:3)))
}
