% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env_is_tabular.R
\name{is_tabular}
\alias{is_tabular}
\title{Confirm an object is table shaped}
\usage{
is_tabular(x)
}
\arguments{
\item{x}{Container object to be checked.}
}
\value{
Logical. Whether all vectors in \code{x} have the same length.
}
\description{
Check whether every vector in an \link{environment} has the same \code{\link[=length]{length()}}
}
